capture dumps on multiple specific processes

+++ procDump.ps1
@@ -51,16 +51,43 @@ if (-not (Test-Path $filePath)) {
 
 #$process = Start-Process -FilePath "dotnet.exe" -PassThru -ArgumentList "tool run ModernUOSchemaGenerator -- ModernUO.sln"
 $toolProcess = Run-Process "dotnet.exe" "tool run ModernUOSchemaGenerator -- ModernUO.sln"
-$dumpProcess = Run-Process "$procDumpFolder\procdump.exe" "-accepteula -ma -s 5 -n 5 $($toolProcess.Id) $procDumpFolder"
 
-while (-not $dumpProcess.HasExited) {
+$dumpProcesses = @()
+$startTime = Get-Date
+while (-not $toolProcess.HasExited) {
+
+    # break after 1 minute as the tool process will likely never exit
+    $currentTime = Get-Date
+    $elapsedTime = $currentTime - $startTime
+
+    if ($elapsedTime.TotalMinutes -ge 1) {
+        Write-Host "Ending, 1 minute has passed."
+        break
+    }
+
+    $commandLines = Get-CimInstance Win32_Process | Select-Object ProcessId, Name, CommandLine
+    foreach ($commandLine in $commandLines) {
+        if ($dumpProcesses -contains $commandLine.ProcessId) {
+            continue
+        }
+
+        if ($commandLine.CommandLine -like "*ModernUOSchemaGenerator.dll*") {
+            $dumpProcesses += $commandLine.ProcessId
+            Write-Host "Attaching to tool process: $($commandLine.ProcessId), Name: $($commandLine.Name), CommandLine: $($commandLine.CommandLine)"
+            $dmp = Run-Process "$procDumpFolder\procdump.exe" "-accepteula -ma -s 5 -n 5 $($commandLine.ProcessId) $procDumpFolder\Tool_PROCESSNAME_YYMMDD_HHMMSS.dmp"
+
+        }
+        if ($commandLine.CommandLine -like "*Microsoft.CodeAnalysis.Workspaces.MSBuild.BuildHost.dll*") {
+            $dumpProcesses += $commandLine.ProcessId
+            Write-Host "Attaching to build host process: $($commandLine.ProcessId), Name: $($commandLine.Name), CommandLine: $($commandLine.CommandLine)"
+            $dmp = Run-Process "$procDumpFolder\procdump.exe" "-accepteula -ma -s 5 -n 5 $($commandLine.ProcessId) $procDumpFolder\BuildHost_PROCESSNAME_YYMMDD_HHMMSS.dmp "
+        }
+    }
+    
     Start-Sleep -Seconds 1
 }
 
-$toolProcess.Kill()
-
-# Start-Sleep -Seconds 30
-
-# $dumpProcess = Run-Process "$procDumpFolder\procdump.exe" "-accepteula -ma -h $($toolProcess.Id) $procDumpFolder"
-# $dumpProcess.WaitForExit()
+if (-not $toolProcess.HasExited) {
+    $toolProcess.Kill()
+}
